+++ src/main/kotlin/si/francebevk/interesnedejavnosti/MainPage.kt
@@ -82,7 +82,7 @@ object MainPage : Action<Chain> {
 
     private fun ActivityRecord.toDTO(isSelected: Boolean) =
         Activity(id, name, description, leader, slots.map { slot ->
-            TimeSlot(translateDay(slot.day), slot.startMinutes.toInt(), slot.endMinutes.toInt())
+            TimeSlot(translateDay(slot.day), slot.startMinutes, slot.endMinutes)
         }, isSelected)
 
 }
